pyproject.toml
[tool.poetry]
name = "fastapi-test"
version = "0.1.0"
description = ""
authors = ["<saddolgov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

alembic = "^1.13.1"
asyncpg = "^0.29.0"
fastapi = "^0.109.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = { extras = ["dotenv", "email"], version = "^2.4.2" }
pydantic-settings = "^2.0.3"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.25"
uvicorn = "^0.27.0"
jinja2 = "^3.1.3"
pydentic = "^0.0.1.dev3"
# social-auth-core = "^4.5.2"
httpx = "^0.26.0"
gunicorn = "^21.2.0"
aiosqlite = "^0.20.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "-v"
asyncio_mode = "auto"
filterwarnings = [
    # Passlib 1.7.4 depends on crypt - https://foss.heptapod.net/python-libs/passlib/-/issues/148
    "ignore:'crypt' is deprecated and slated for removal",
]
markers = ["pytest.mark.asyncio"]
minversion = "6.0"
testpaths = ["app/tests"]

[tool.ruff]
target-version = "py312"
# pycodestyle, pyflakes, isort, pylint, pyupgrade
select = ["E", "W", "F", "I", "PL", "UP"]
ignore = ["E501"]

